Add Network and Remove Unused Options
I added the ability to add a network. This lets you add a range of ip
addresses. To do so select network from the engagement_type options and
add the following to network.

`{"base": "192.168.1", "sets": [1,2,3]}`

I wanted this because of the OSCP exercises. Some of them are several
VMs. This lets me create notes for each.

I also, the files that were removes were done som becaus I did not feel
it was nessesary. I also have those in the nots itself.