Add Directories and Files
I added three commenly used directories and in each a README.txt file.
These README files are empty and just so that git will pick up the
changes and allow me to creat the directories. The README is a .txt and
not a .md because I do not want obsidian to pick it up. I also added a
user and password file to.